Fix for #402: When mask is "comma_separator.2", Input with numeric mask won't bind back correct value to model when value of field is blank
This is a fix for this issue:

Releated to Issue #380, when the user enters a numeric value matching the mask the component model is updated with that value, however when the user empties the input so that the value should be "" it is instead 0.00.

This bug was fixed in #380 but is still a problem when the mask is "comma_separator.2"

<input formControlName="TransportDollarsPerMile" class="form-control" id="transportDollarsPerMileId" type="text" placeholder="$" prefix="$" mask="comma_separator.2" />

When I clear out the input box, group.get('TransportDollarsPerMile') is 0.00. It should be null or "" so my angular required field validator can work.

I haven't tested this locally since I wasn't able to get the dev environment up so be sure to test it!